Update everything to typescript

+++ api/src/internal/models.go
@@ -1,32 +1,32 @@
 package internal
 
 type DBPredictedYield struct {
-	YieldKW  int64
-	TimeUnix int32
-	ErrLow   int64
-	ErrHigh  int64
+	YieldKW  int
+	TimeUnix int64
+	ErrLow   int
+	ErrHigh  int
 }
 
 type DBActualYield struct {
-	YieldKW  int64
-	TimeUnix int32
+	YieldKW  int
+	TimeUnix int64
 }
 
 type DBPredictedLocalisedYield struct {
 	LocationID string
-	YieldKW    int64
-	ErrLow     int64
-	ErrHigh    int64
+	YieldKW    int
+	ErrLow     int
+	ErrHigh    int
 }
 
 type DBActualLocalisedYield struct {
 	LocationID string
-	YieldKW    int64
+	YieldKW    int
 }
 
 type DatabaseService interface {
 	GetPredictedYieldsForLocation(locID string) ([]DBPredictedYield, error)
 	GetActualYieldsForLocation(locID string) ([]DBActualYield, error)
-	GetPredictedYieldForLocations(locIDs []string, timeUnix int32) ([]DBPredictedLocalisedYield, error)
-	GetActualYieldForLocations(locIDs []string, timeUnix int32) ([]DBActualLocalisedYield, error)
+	GetPredictedYieldForLocations(locIDs []string, timeUnix int64) ([]DBPredictedLocalisedYield, error)
+	GetActualYieldForLocations(locIDs []string, timeUnix int64) ([]DBActualLocalisedYield, error)
 }
